feat:add-validate-to-users

+++ src/modules/users/services/users.service.ts
@@ -1,8 +1,10 @@
 import { PrismaService } from '@modules/prisma/infra/database/prisma.service';
-import { Injectable } from '@nestjs/common';
-import { CreateUserDto, UpdateUserDto, User } from '../dtos/userDTO';
+import { ConflictException, Injectable } from '@nestjs/common';
+import { CreateUserDto, UpdateUserDto } from '../dtos/userDTO';
 import { UsersRepository } from '../repositories/implementations/prisma-users-repository';
 import { IUsersRepository } from '../repositories/IUsers-repository';
+import { hash } from 'bcryptjs';
+import { User } from '@prisma/client';
 
 
 @Injectable()
@@ -10,8 +12,13 @@ export class UsersService {
   constructor(private readonly usersRepository: IUsersRepository) {}
 
   async create(createUserDto: CreateUserDto): Promise<User> {
+    const userExist = await this.usersRepository.findByEmailORCpf(createUserDto.email, createUserDto.cpf)
+    if(userExist)
+      throw new ConflictException('usuário já existente');
+
+    const passwordHash = await hash(createUserDto.password, 8);
+    createUserDto.password = passwordHash;
     const user = await this.usersRepository.create(createUserDto);
-    console.log(user);
     return user;
   }
 
@@ -26,12 +33,19 @@ export class UsersService {
   }
 
   async update(id: string, updateUserDto: UpdateUserDto): Promise<User> {
+    const userExist = await this.usersRepository.findById(id)
+
+    if(userExist)
+      throw new ConflictException('usuário já existente');
 
     const user = await this.usersRepository.update(id, updateUserDto)
     return user;
   }
 
   async remove(id: string): Promise<void> {
+    const userExist = await this.usersRepository.findById(id)
+    if(userExist)
+      throw new ConflictException('usuário já existente');
     await this.usersRepository.deleteById(id)
   }
 }
